moved logorf to organisms
